+++ src/types.ts
@@ -95,31 +95,6 @@ export type TsonType<
 	TSerializedType extends SerializedType,
 > = TsonTypeTester & TsonTransformer<TValue, TSerializedType>;
 
-export type TsonAsyncIndex = TsonBranded<number, "AsyncRegistered">;
-export interface TsonTransformerSerializeDeserializeAsync<TValue> {
-	async: true;
-	/**
-	 * From JSON-serializable value
-	 */
-	deserialize: (v: TsonAsyncIndex) => TValue;
-
-	/**
-	 * The key to use when serialized
-	 */
-	key: string;
-	/**
-	 * JSON-serializable value
-	 */
-	serialize: (
-		v: TValue,
-		register: (thing: TValue) => TsonAsyncIndex,
-	) => TsonAsyncIndex;
-}
-
-export interface TsonAsyncType<TValue>
-	extends TsonTransformerSerializeDeserializeAsync<TValue>,
-		TsonTypeTesterCustom {}
-
 export interface TsonOptions {
 	/**
 	 * The nonce function every time we start serializing a new object
@@ -131,19 +106,6 @@ export interface TsonOptions {
 	 * The list of types to use
 	 */
 	types: (TsonType<any, any> | TsonType<any, never>)[];
-}
-
-export interface TsonAsyncOptions {
-	/**
-	 * The nonce function every time we start serializing a new object
-	 * Should return a unique value every time it's called
-	 * @default `${crypto.randomUUID} if available, otherwise a random string generated by Math.random`
-	 */
-	nonce?: () => number | string;
-	/**
-	 * The list of types to use
-	 */
-	types: (TsonAsyncType<any> | TsonType<any, any> | TsonType<any, never>)[];
 }
 
 export const serialized = Symbol("serialized");
